+++ custom_components/alarmdotcom/button.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass
 from inspect import isawaitable
 from typing import TYPE_CHECKING, Any, Generic
 
-import pyalarmdotcomajax as pyadc
+import _pyalarmdotcomajax as pyadc
 from homeassistant.components.button import ButtonEntity, ButtonEntityDescription
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import EntityCategory